Refactor ConditionTraceWriter
